Migrate 'property' test
